requirements.txt
fastapi
jinja2
python-multipart
uvicorn
python-dotenv
httpx
